+++ AppCore/Models/PaginationConfiguration.py
@@ -1,0 +1,4 @@
+class PaginationConfiguration:
+    def __init__(self, page: int, page_size: int):
+        self.page = page
+        self.page_size = page_size