+++ AppUI/Clients/swu_db_com/SWUDBAPIRemoteClient.py
@@ -1,7 +1,9 @@
+from typing import Optional
+
 from AppCore.Data.APIClientProtocol import (APIClientProtocol,
-                                            APIClientSearchCallback,
-                                            NetworkerProtocol)
-from AppCore.Models import SearchConfiguration
+                                            APIClientSearchCallback)
+from AppCore.Models import PaginationConfiguration, SearchConfiguration
+from AppCore.Network import NetworkerProtocol
 
 from ..SWUCardSearchConfiguration import SWUCardSearchConfiguration
 from .SearchRequest import SearchRequest
@@ -22,6 +24,10 @@ class SWUDBAPIRemoteClient(APIClientProtocol):
     def site_source_url(self) -> str:
         return "https://www.swu-db.com/"
 
-    def search(self, search_configuration: SearchConfiguration, callback: APIClientSearchCallback):
+    def search(self, 
+               search_configuration: SearchConfiguration,
+               pagination_configuration: Optional[PaginationConfiguration],
+               callback: APIClientSearchCallback):
         swu_search_config = SWUCardSearchConfiguration.from_search_configuration(search_configuration)
+        
         self._networker.load(SearchRequest(swu_search_config), callback)