+++ AppUI/Clients/swu_db_com/SearchRequest.py
@@ -4,6 +4,7 @@ from urllib.request import Request
 
 from AppCore.Models import TradingCard
 from AppCore.Network import NetworkRequestProtocol
+from AppCore.Data import APIClientSearchResponse
 
 from AppCore.Models.CardAspect import CardAspect
 from AppCore.Models.CardType import CardType
@@ -11,10 +12,11 @@ from ..SWUCardSearchConfiguration import SWUCardSearchConfiguration
 from .SWUTradingCard import SWUTradingCard
 
 # https://api.swu-db.com/cards/search?q=type:leader%20name:luke
-class SearchRequest(NetworkRequestProtocol[List[TradingCard]]):
+class SearchRequest(NetworkRequestProtocol[APIClientSearchResponse]):
         SWUDB_API_ENDPOINT = 'https://api.swu-db.com/cards/search'
         
-        def __init__(self, search_configuration: SWUCardSearchConfiguration):
+        def __init__(self, 
+                     search_configuration: SWUCardSearchConfiguration):
             self.search_configuration = search_configuration
         
         def __eq__(self, other):  # type: ignore
@@ -44,12 +46,12 @@ class SearchRequest(NetworkRequestProtocol[List[TradingCard]]):
             print(url)
             return Request(url)
         
-        def response(self, json: Dict[str, Any]) -> List[TradingCard]:
+        def response(self, json: Dict[str, Any]) -> APIClientSearchResponse:
             result_list: List[TradingCard] = []
             for i in json['data']:
                 swu_card = SWUTradingCard.from_swudb_response(i)
                 result_list.append(swu_card)
-            return result_list
+            return APIClientSearchResponse(result_list)
         
         @staticmethod
         def _aspect_query_mapping() -> Dict['CardAspect', str]: